requirements.txt
# Core Framework
pydantic==2.5.2
pyyaml==6.0.1
jsonpickle==3.0.2
click==8.1.7

# New CLI (Rich + Typer)
rich==13.7.0
typer==0.12.0

# State Management (PostgreSQL only)
psycopg2-binary==2.9.9
sqlalchemy==1.4.50

# Data I/O
kafka-python==2.0.2
boto3==1.34.7
pandas==1.3.5
pyarrow==14.0.2

# PyTorch Ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0

# API Framework (FastAPI)
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# HTTP & Auth
httpx==0.23.3
requests==2.31.0

# Kubernetes & Docker
kubernetes==28.1.0
docker==7.0.0

# Metrics & Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
shortuuid==1.0.11
retrying==1.3.4

# Logging & Config
python-dateutil==2.8.2
pytz==2024.1

# Vision-specific
Pillow>=10.0.0
opencv-python>=4.8.0
easyocr>=1.7.0
paddleocr>=2.7.0
uform>=1.3.0

# Audio-specific
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.2
pyctcdecode==0.5.0
whispercpp==0.0.17
faster-whisper==1.0.2

# Text/NLP-specific
sentencepiece>=0.1.99
sacremoses>=0.0.53
tokenizers>=0.15.0

# Optimization & Quantization
bitsandbytes>=0.41.0
accelerate>=0.25.0
optimum>=1.16.0

# Dev Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.8.0
coverage==7.3.4
